Refactored Mongoose code to ts
